# cmd/settlr/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Settlr API",
    "version": "1.0.0",
    "description": "Free/open-source Splitwise alternative. Integer cents ledger, Argon2id, JWT refresh rotation, cursor pagination."
  },
  "servers": [
    {
      "url": "https://settlrapi.theswissknife.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/friends": {
      "get": {
        "summary": "List friends",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/friends/requests": {
      "get": {
        "summary": "Pending requests",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/friends/{id}/request": {
      "post": {
        "summary": "Send request",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/friends/{id}/accept": {
      "post": {
        "summary": "Accept",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/friends/{id}/reject": {
      "post": {
        "summary": "Reject",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/friends/{id}": {
      "delete": {
        "summary": "Remove friend",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/friends/{id}/block": {
      "post": {
        "summary": "Block",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/notifications": {
      "get": {
        "summary": "Notifications",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/notifications/{id}/read": {
      "post": {
        "summary": "Mark read",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "summary": "Mark all read",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}